requirements.txt
# Core Flask Framework
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Face Recognition & Image Processing
face_recognition==1.3.0
face-recognition-models==0.3.0
dlib==19.24.2
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4

# Utilities
setuptools==68.2.2
click==8.2.1
